-CVARS should utilize the asInt() member function when implicitly or explicitly being cast as an integer type. -asInt() provides rounding which avoids truncating the value of the CVAR when converting from its internal storage format of float to an integer type. -When assigning a value of 1 in a variable of type float, the float's stored value is typically something like 0.9999999972 or 1.00000000018 due to the imprecise nature of floating point variables.  When naively casting the former to an integer type, the value would be truncated to 0 instead of 1.
SVN r2320 (0.6-dev)

+++ common/p_interaction.cpp
@@ -1204,7 +1204,7 @@ void P_KillMobj(AActor *source, AActor *target, AActor *inflictor, bool joinkill
 	{
 		// [Toke] Better sv_fraglimit
 		if (sv_gametype == GM_DM && sv_fraglimit &&
-            splayer->fragcount >= (int)sv_fraglimit && !sv_fragexitswitch)
+            splayer->fragcount >= sv_fraglimit && !sv_fragexitswitch)
 		{
             // [ML] 04/4/06: Added !sv_fragexitswitch
             SV_BroadcastPrintf(
@@ -1220,7 +1220,7 @@ void P_KillMobj(AActor *source, AActor *target, AActor *inflictor, bool joinkill
 		{
 			for (size_t i = 0; i < NUMFLAGS; i++)
 			{
-				if (TEAMpoints[i] >= (int)sv_fraglimit)
+				if (TEAMpoints[i] >= sv_fraglimit)
 				{
 					SV_BroadcastPrintf(
                         PRINT_HIGH,
